+++ docs/source/conf.py
@@ -18,7 +18,7 @@ author = "Tom Strange"
 
 extensions = [
     "sphinx.ext.autodoc",
-    "sphinx.ext.napoleon",
+    "numpydoc",
     "sphinx.ext.viewcode",
     "sphinx.ext.intersphinx",
     "sphinx_copybutton",
@@ -26,12 +26,8 @@ extensions = [
     "myst_parser",
 ]
 
-# autodoc_typehints = "none"
-napoleon_numpy_docstring = True
-napoleon_use_admonition_for_references = True
-napoleon_use_param = False
-napoleon_use_rtype = False
-napoleon_use_keyword = False
+numpydoc_xref_param_type = True
+numpydoc_class_members_toctree = False
 
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]