+++ ingredient_parser/en/_utils.py
@@ -48,7 +48,8 @@ MISINTERPRETED_UNITS = [
     "shakes",
     "tin",  # tera-inch
     "tins",
-    "fat",  # femto-technical-atmosphere
+    "unit",  # micronit (micro netwon inch)
+    "unitsfat",  # femto-technical-atmosphere
 ]
 
 # List of unit replacements so that these units get converted to the correct pint