+++ train/featuresearch.py
@@ -2,6 +2,7 @@
 
 import argparse
 import concurrent.futures as cf
+import logging
 import os
 import time
 from datetime import timedelta
@@ -19,6 +20,8 @@ from .training_utils import (
     evaluate,
     load_datasets,
 )
+
+logger = logging.getLogger(__name__)
 
 DISCARDED_FEATURES = {
     0: [],
@@ -71,6 +74,7 @@ def train_model_feature_search(
     save_model: str,
     seed: int,
     keep_model: bool,
+    combine_name_labels: bool,
 ) -> dict:
     """Train model using selected features returning model performance statistics,
     model parameters and elapsed training time.
@@ -90,6 +94,8 @@ def train_model_feature_search(
         testing sets.
     keep_model : bool
         If True, keep model after evaluation, otherwise delete it.
+    combine_name_labels : bool, optional
+        If True, combine all NAME labels into a single NAME label.
 
     Returns
     -------
@@ -154,7 +160,7 @@ def train_model_feature_search(
     tagger = pycrfsuite.Tagger()  # type: ignore
     tagger.open(str(save_model_path))
     labels_pred = [tagger.tag(X) for X in features_test]
-    stats = evaluate(labels_pred, truth_test, seed)
+    stats = evaluate(labels_pred, truth_test, seed, combine_name_labels)
 
     if not keep_model:
         save_model_path.unlink(missing_ok=True)
@@ -176,7 +182,13 @@ def feature_search(args: argparse.Namespace):
     args : argparse.Namespace
         Feature search configuration
     """
-    vectors = load_datasets(args.database, args.table, args.datasets)
+    vectors = load_datasets(
+        args.database,
+        args.table,
+        args.datasets,
+        discard_other=True,
+        combine_name_labels=args.combine_name_labels,
+    )
 
     if args.save_model is None:
         save_model = DEFAULT_MODEL_LOCATION
@@ -192,19 +204,21 @@ def feature_search(args: argparse.Namespace):
             save_model,
             args.seed,
             args.keep_models,
+            args.combine_name_labels,
         ]
         argument_sets.append(arguments)
 
-    print(f"[INFO] Grid search over {len(argument_sets)} feature sets.")
-    print(f"[INFO] {args.seed} is the random seed used for the train/test split.")
+    logger.info(f"Grid search over {len(argument_sets)} feature sets.")
+    logger.info(f"{args.seed} is the random seed used for the train/test split.")
 
-    eval_results = []
     with cf.ProcessPoolExecutor(max_workers=args.processes) as executor:
         futures = [
             executor.submit(train_model_feature_search, *a) for a in argument_sets
         ]
-        for future in tqdm(cf.as_completed(futures), total=len(futures)):
-            eval_results.append(future.result())
+        eval_results = [
+            future.result()
+            for future in tqdm(cf.as_completed(futures), total=len(futures))
+        ]
 
     # Sort with highest sentence accuracy first
     eval_results = sorted(