+++ train/viz/performance_plot.py
@@ -25,12 +25,10 @@ def load_data() -> tuple[tuple[str], list[float], list[float]]:
         sentence is a list of floats
         word is a list of floats
     """
-    data = []
     with open("train/performance_history.csv", "r") as f:
         reader = csv.reader(f)
-        next(reader)
-        for row in reader:
-            data.append(row)
+        # Skip first row because it contains column headings
+        data = list(reader)[1:]
 
     releases, sentence, word = zip(*data)
     sentence = [float(s) for s in sentence]
@@ -43,7 +41,7 @@ def main():
     """Plot figure and save to docs/source/_static folder"""
     releases, sentence, word = load_data()
 
-    fig, ax = plt.subplots(figsize=(12, 5), layout="constrained")
+    fig, ax = plt.subplots(figsize=(12, 6), layout="constrained")
 
     x = np.arange(0, 2 * len(releases), 2)  # the label locations
     width = 0.95  # the width of the bars
@@ -55,7 +53,7 @@ def main():
     ax.bar_label(rects, padding=-40, fontsize=12, rotation=90, fmt="%.2f")
 
     offset = width
-    rects = ax.bar(x + offset, word, width, label="Word accuracy", color="#9f4e19")
+    rects = ax.bar(x + offset, word, width, label="Token accuracy", color="#9f4e19")
     ax.bar_label(rects, padding=-40, fontsize=12, rotation=90, fmt="%.2f")
 
     ax.spines["top"].set_visible(False)
@@ -68,7 +66,7 @@ def main():
     ax.set_ylabel("Accuracy (%)", fontsize=18)
     ax.set_xticks(x + width / 2, releases, fontsize=12, rotation=45)
     ax.legend(loc="upper left", ncols=1, fontsize=16)
-    ax.set_ylim(80, 102)
+    ax.set_ylim(82, 101)
     ax.set_xlim(-0.7, 2 * len(releases) - 0.3)
     fig.savefig("docs/source/_static/performance-history.svg")